+++ fili-core/src/main/java/com/yahoo/bard/webservice/web/handlers/RequestContext.java
@@ -26,7 +26,7 @@ public class RequestContext {
     /**
      * Build a context for a request.
      *
-     * @param containerRequestContext  context from the http request object
+     * @param containerRequestContext  context from the http request object (nullabe in embedded contexts)
      * @param readCache  true if the cache should be checked for a response
      */
     public RequestContext(ContainerRequestContext containerRequestContext, boolean readCache) {